feat: changed the recycler view of ProviderRequestFragment to make it dinamic with firebase

+++ app/src/main/java/com/ar/of_pro/entities/ServiceProvider.kt
@@ -3,15 +3,15 @@ package com.ar.of_pro.entities
 import android.os.Parcel
 import android.os.Parcelable
 
-class ServiceProvider(name: String?, bidAmount: Float?, calification: Double?): Parcelable {
+class ServiceProvider(name: String?, bidAmount: Float?, calification: Float?): Parcelable {
     var name: String = ""
     var bidAmount: Float = 0f
-    var calification: Double = 0.0
+    var calification: Float = 0f
 
     constructor(parcel: Parcel) : this(
         parcel.readString(),
         parcel.readFloat(),
-        parcel.readDouble()
+        parcel.readFloat()
     )
 
     init {
@@ -22,7 +22,7 @@ class ServiceProvider(name: String?, bidAmount: Float?, calification: Double?): 
     override fun writeToParcel(parcel: Parcel, flags: Int) {
         parcel.writeString(name)
         parcel.writeFloat(bidAmount)
-        parcel.writeDouble(calification)
+        parcel.writeFloat(calification)
     }
 
     override fun describeContents(): Int {
